+++ src/MapPublisher.cpp
@@ -1,7 +1,7 @@
 #include "MapPublisher.h"
 
 
-void MapPublisher::PublishPose(Matrix4f pose, int type)
+void MapPublisher::PublishPose(Matrix4d pose, int type)
 {
 
 //    static tf::TransformBroadcaster mTfBr;
@@ -10,7 +10,7 @@ void MapPublisher::PublishPose(Matrix4f pose, int type)
 //    mTfBr.sendTransform(tf::StampedTransform(tfT,ros::Time::now(), "/CamLoc/World", "/CamLoc/Camera"));
 
     visualization_msgs::Marker mCurrentCamera;
-    vector<Matrix4f>* Poses;
+    vector<Matrix4d>* Poses;
     switch (type){
     case 1:
         mCurrentCamera = mOdoCamera;
@@ -32,13 +32,13 @@ void MapPublisher::PublishPose(Matrix4f pose, int type)
     mCurrentCamera.points.clear();
     float d = fCameraSize;
     //Camera is a pyramid. Define in camera coordinate system
-    cv::Mat o = (cv::Mat_<float>(4,1) << 0, 0, 0, 1);
-    cv::Mat p1 = (cv::Mat_<float>(4,1) << d, d*0.8, d*0.5, 1);
-    cv::Mat p2 = (cv::Mat_<float>(4,1) << d, -d*0.8, d*0.5, 1);
-    cv::Mat p3 = (cv::Mat_<float>(4,1) << -d, -d*0.8, d*0.5, 1);
-    cv::Mat p4 = (cv::Mat_<float>(4,1) << -d, d*0.8, d*0.5, 1);
+    cv::Mat o = (cv::Mat_<double>(4,1) << 0, 0, 0, 1);
+    cv::Mat p1 = (cv::Mat_<double>(4,1) << d, d*0.8, d*0.5, 1);
+    cv::Mat p2 = (cv::Mat_<double>(4,1) << d, -d*0.8, d*0.5, 1);
+    cv::Mat p3 = (cv::Mat_<double>(4,1) << -d, -d*0.8, d*0.5, 1);
+    cv::Mat p4 = (cv::Mat_<double>(4,1) << -d, d*0.8, d*0.5, 1);
 
-    for(vector<Matrix4f>::iterator vit=Poses->begin(), vend=Poses->end(); vit!=vend; vit++){
+    for(vector<Matrix4d>::iterator vit=Poses->begin(), vend=Poses->end(); vit!=vend; vit++){
         cv::Mat Twc;
         cv::eigen2cv((*vit),Twc);
         cv::Mat ow = Twc*o;
@@ -48,21 +48,21 @@ void MapPublisher::PublishPose(Matrix4f pose, int type)
         cv::Mat p4w = Twc*p4;
 
         geometry_msgs::Point msgs_o,msgs_p1, msgs_p2, msgs_p3, msgs_p4;
-        msgs_o.x=ow.at<float>(0);
-        msgs_o.y=ow.at<float>(1);
-        msgs_o.z=ow.at<float>(2);
-        msgs_p1.x=p1w.at<float>(0);
-        msgs_p1.y=p1w.at<float>(1);
-        msgs_p1.z=p1w.at<float>(2);
-        msgs_p2.x=p2w.at<float>(0);
-        msgs_p2.y=p2w.at<float>(1);
-        msgs_p2.z=p2w.at<float>(2);
-        msgs_p3.x=p3w.at<float>(0);
-        msgs_p3.y=p3w.at<float>(1);
-        msgs_p3.z=p3w.at<float>(2);
-        msgs_p4.x=p4w.at<float>(0);
-        msgs_p4.y=p4w.at<float>(1);
-        msgs_p4.z=p4w.at<float>(2);
+        msgs_o.x=ow.at<double>(0);
+        msgs_o.y=ow.at<double>(1);
+        msgs_o.z=ow.at<double>(2);
+        msgs_p1.x=p1w.at<double>(0);
+        msgs_p1.y=p1w.at<double>(1);
+        msgs_p1.z=p1w.at<double>(2);
+        msgs_p2.x=p2w.at<double>(0);
+        msgs_p2.y=p2w.at<double>(1);
+        msgs_p2.z=p2w.at<double>(2);
+        msgs_p3.x=p3w.at<double>(0);
+        msgs_p3.y=p3w.at<double>(1);
+        msgs_p3.z=p3w.at<double>(2);
+        msgs_p4.x=p4w.at<double>(0);
+        msgs_p4.y=p4w.at<double>(1);
+        msgs_p4.z=p4w.at<double>(2);
 
         mCurrentCamera.points.push_back(msgs_o);
         mCurrentCamera.points.push_back(msgs_p1);
